Solve errors when no standards were loaded
If any standard file has been loaded in the current session (because the path in the options file is incorrect or because no standards were loaded manually) and you try to fit any region which calls a standard, the program will crash and a waiting dialog will appear permanently. Matlab needs to be killed from the task manager.